зачатки FLUX концепции
